Add files via upload
Xmss tree created with n signatures from a SEED. Smaller trees are now creatable and used to derive addresses from the same SEED within the xmss class wrapper. They are stored and can be signed for in the same way as the main address. Given enough time they can be restored from blockchain like an HD wallet in bitcoin.