Optimized memory and disk usage, Block compression added
Block could be stored into disk, either with bz2 or zlib
compression.

Now, complete chain will not be maintained in the memory,
only chain of current epoch will be available.

Leveldb is now maintaing block metadata such as offset
from where the block could be found in the file and the
size required to read the block from the file.