fix mining timestamp attack vulnerability
Timestamp checks were previously:
1) confirm new block > 120s into future compared with local node ntp-time
2) check that new block timestamp < previous block timestamp

It is possible for a modified client to mine blocks which are 120s into the future. In that situation untampered nodes would be unable to mine blocks on top of the block with a fraudulent timestamp as rule 2) would render such a block invalid. 

Following this strategy it is possible for a miner to win every block by faking the timestamps perpetually into the future. Removing this condition fixes the vulnerability. We can discuss a fix which allows variability in the timestamp in the blockheader in such a way that difficulty is unaffected, whilst not exposing a timestamp attack as detailed above.